# perf: remove unnecessary join on UserAccounts