#1712 removes the untracked user filtering for sets with over 100 players but the join on UserAccounts was still there so this removes that as well.

**Benchmarks**
These samples were collected on game ID 1446 with the local MySQL cache disabled. Results are the combine 2 calls for hardcore and softcore

  | Before | After
-- | -- | --
1 | 1105ms | 507ms
2 | 1180ms| 501ms
3 | 1117ms | 505ms
4 | 1135ms | 493ms
5 | 1119ms | 499ms
Average | 1131ms | 501ms

Speedup Factor: ~2.26x